==> code/02_06_get_seq_id_accessions.ipynb <==
import pandas as pd

seq_assembly_df = pd.read_csv('../data/interim/seq_assemblies.csv',
                          names=['protein_accession', 'seq_id', 'seq', 'assembly'])


unique_seq_id_accessions = seq_assembly_df[['protein_accession', 'seq_id']].drop_duplicates()
del seq_assembly_df

unique_seq_id_accessions.to_parquet('../data/interim/seq_id_accessions.pq', index=False)

unique_seq_id_accessions = pd.read_parquet('../data/interim/seq_id_accessions.pq')

unique_seq_id_accessions.shape
