==> code/02_08_get_protein_go_terms.ipynb <==
import pandas as pd
import os
import re
from joblib import Parallel, delayed
from tqdm import tqdm

def get_go_processes(in_file):
    gff_df = pd.read_table('../data/genome_downloads/gff/' + in_file,
                           names=['seqid', 'source', 'feature', 'start', 'end',
                                  'score', 'strand', 'frame', 'attributes'],
                           comment='#')
    filtered_gff_df = gff_df[gff_df['attributes'].str.contains('go_process', na=False) &
                             ~gff_df['attributes'].str.contains('pseudo', na=False)]
    assembly = in_file.split('.')[0]
    with open('../data/interim/go_process_temp/' + assembly + '.csv', 'w') as f:
        for attr in filtered_gff_df['attributes']:
            prot_id = re.search('protein_id=([^;]+)', attr).group(1)
            go_processes = re.search('go_process=([^;]+)', attr).group(1)
            for process in go_processes.split(','):
                process = process.split('|')[0]
                print(prot_id + ',' + process, file=f)

os.mkdir('../data/interim/go_process_temp/')
gff_files = [x for x in os.listdir('../data/genome_downloads/gff') if '.gff' in x]
print('Extracting go processes')
_ = Parallel(n_jobs=48)(delayed(get_go_processes)(file) for file in tqdm(gff_files))
os.system('cat ../data/interim/go_process_temp/*.csv > ../data/interim/go_processs_temp.csv')
os.system('rm -r ../data/interim/go_process_temp/')

print('Getting unique go processes')
unique_go_processes = (pd.read_csv('../data/interim/go_processs_temp.csv',
                                   names=['product_accession', 'go_process'])
                       .drop_duplicates()
                       .set_index('product_accession'))

os.system('rm ../data/interim/go_processs_temp.csv')

print('Merging seq ids with go processes')
unique_seq_id_accessions = (pd.read_parquet('../data/interim/seq_id_accessions.pq')
                            .set_index('protein_accession'))
unique_seq_ids_go_process = (unique_go_processes.merge(unique_seq_id_accessions, how='inner',
                                                           left_index=True, right_index=True)
                             [['seq_id', 'go_process']]
                             .drop_duplicates(ignore_index=True)) # so new indeces are labeled 0, 1, ...

print('Proteins with a go process', unique_seq_ids_go_process['seq_id'].nunique())

print('Top go processes')
unique_seq_ids_go_process['go_process'].value_counts().head(10)

print('Number of go processes', unique_seq_ids_go_process['go_process'].nunique())

unique_seq_ids_go_process.to_parquet('../data/interim/seq_id_go_processes.pq', index=False)
