==> code/03_04_get_defense_model_seqs.ipynb <==
import pandas as pd
import os
import glob
import xml.etree.ElementTree as ET
import networkx as nx
from tqdm import tqdm
from core import read_mmseqs_results

def get_xml_file_paths(top_directory):
    pattern = f"{top_directory}/**/*.xml"
    file_paths = glob.glob(pattern, recursive=True)    
    return file_paths

def get_exhchangable_profiles(file_path):
    model = os.path.basename(file_path).split('.')[0]
    tree = ET.parse(os.path.expanduser(file_path))
    root = tree.getroot()
    data = []
    exchangeable_names = []
    parent_names = []
    for gene in root.findall('.//gene'):
        gene_name = gene.attrib.get('name', '')
        exchangeables = gene.findall('.//exchangeables')
        if exchangeables:
            parent_names.append(gene_name)
            for exch in exchangeables:
                exch_genes = exch.findall('.//gene')
                for sub_gene in exch_genes:
                    exch_name = sub_gene.attrib.get('name', '')
                    data.append({'model': model, 'parent_profile': gene_name, 'profile': exch_name})
                    exchangeable_names.append(exch_name)
        data.append({'model': model, 'parent_profile': gene_name, 'profile': gene_name})
    df = pd.DataFrame(data)
    rm_exchangable = [x for x in exchangeable_names if x not in parent_names]
    df = df[~df['parent_profile'].isin(rm_exchangable)].reset_index(drop=True)
    return df

## Part 1: Get defense genes to use for modeling

defense_finder_genes_df = pd.read_parquet('../data/interim/defense_finder_genes_genomes.pq')
defense_homolog_df = pd.read_parquet('../data/interim/defense_finder_homologs_profile_names.pq')
select_cluster_df = pd.read_parquet('../data3/interim/refseq_cover_clusters.pq')

true_defense_seq_ids = defense_finder_genes_df['seq_id'].drop_duplicates()
print(len(true_defense_seq_ids))
defense_homolog_seq_ids = defense_homolog_df['seq_id'].drop_duplicates()
print(len(defense_homolog_seq_ids))

%%time
all_seqs = pd.read_parquet('../data/interim/refseq_seq_ids.pq')

minimal_true_defense_seq_ids = true_defense_seq_ids[true_defense_seq_ids
                                                    .isin(select_cluster_df['cluster_id'])]

len(minimal_true_defense_seq_ids)

select_defense_finder_genes = (defense_finder_genes_df[defense_finder_genes_df['seq_id']
                                                       .isin(minimal_true_defense_seq_ids)]
                               .groupby('seq_id')
                               .sample(n=1, random_state=7))

select_defense_finder_genes['gene_name'].value_counts().head(20)

## Part2: Get HMM clusters

### Edge type: exchangable

definition_file_paths = get_xml_file_paths(os.path.expanduser('~/defense-finder-models/definitions/'))

len(definition_file_paths)

gene_group_list = []
for f in tqdm(definition_file_paths):
    gene_group_list.append(get_exhchangable_profiles(f))
gene_group_df = pd.concat(gene_group_list)

gene_group_edges = (gene_group_df.merge(gene_group_df, how='inner', on='parent_profile')
                    .query('profile_x != profile_y')
                    [['profile_x', 'profile_y']]
                    .drop_duplicates()
                    .rename(columns={'profile_x': 'gene_name_x', 
                                     'profile_y': 'gene_name_y'}))

len(gene_group_edges)

all_genes = gene_group_df['profile'].drop_duplicates()
len(all_genes)

G = nx.from_pandas_edgelist(gene_group_edges, 'gene_name_x', 'gene_name_y')
nodes = list(G.nodes)
missing_nodes = all_genes[~all_genes.isin(nodes)].to_list()
G.add_nodes_from(missing_nodes)
print('# Edges:', G.number_of_edges())
print('# Nodes:', G.number_of_nodes())
components = nx.connected_components(G)
component_list = []
for i, comp in enumerate(components):
    subgraph = G.subgraph(comp)
    max_degree_node = max(subgraph, key=lambda node: subgraph.degree(node))
    for name in comp:
        component_list.append({'gene_name': name, 
                               'component': i, 
                               'component_name': max_degree_node})
print('# Components:', i)
exchangable_component_df = pd.DataFrame(component_list)
exchangable_component_df['component_name'].value_counts().head(20)

exchangable_component_df

exchangable_component_df = exchangable_component_df.rename(columns={'component': 'defense_exchangable_num',
                                                                    'component_name': 'defense_exchangable_name'})

defense_seqs = all_seqs[all_seqs['seq_id'].isin(true_defense_seq_ids)]

merged_select_genes = (select_defense_finder_genes
                       .merge(exchangable_component_df, how='inner',
                              on='gene_name')
                       .merge(defense_seqs, how='inner',
                              on='seq_id'))

defense_seq_f = '../data3/interim/select_defense_seqs.faa'

with open(defense_seq_f, 'w') as f:
    for _, row in merged_select_genes.iterrows():
        print('>' + row['seq_id'], file=f)
        print(row['seq'], file=f)

defense_edges = (merged_select_genes[['seq_id', 'defense_exchangable_name']]
                 .merge(merged_select_genes[['seq_id', 'defense_exchangable_name']], 
                        how='inner', on='defense_exchangable_name'))

defense_edges.to_parquet('../data3/interim/defense_exchangable_edges.pq', index=False)

merged_select_genes.to_parquet('../data3/interim/defense_finder_model_seq_info.pq', index=False)
