==> code/04_00_get_defense_go_terms.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

go_df = pd.read_parquet('../data/interim/seq_id_go_processes.pq')
true_defense_genes = pd.read_parquet('../data/interim/defense_finder_true_defense_seq_ids.pq')

seq_id_names = pd.read_parquet('../data/interim/seq_id_names.pq')

If a gene **might** (p>0.1) be defenseive based on its GO process, then don't include it in the negative set. i.e., if p(defense|GO process)>0.1, remove gene from the negative list

### GO term filter

defense_go_df = go_df[go_df['seq_id'].isin(true_defense_genes['seq_id'])]

defense_go_stats = (defense_go_df['go_process'].value_counts()
                     .reset_index(name='count')
                     .rename(columns={'index': 'go_process'}))

baseline_go_stats = (go_df['go_process'].value_counts()
                     .reset_index(name='count')
                     .rename(columns={'index': 'go_process'}))

merged_go_stats = (defense_go_stats.merge(baseline_go_stats, how='inner', on='go_process', 
                                          suffixes=('_defense', '_baseline')))
merged_go_stats['defense_percent'] = merged_go_stats['count_defense']/merged_go_stats['count_baseline']

merged_go_stats.sort_values('defense_percent', ascending=False).head(20)

defense_go_terms = merged_go_stats.loc[merged_go_stats['defense_percent'] > 0.1, 'go_process'].to_list()

We'll filter out genes that are members of the following  pathways

defense_go_terms

defense_go_term_seq_ids = go_df.loc[go_df['go_process'].isin(defense_go_terms), 
                                    'seq_id'].unique()

### T/A

Due to their frequent misannotation in stress response or persister cell formation, we will remove obvious TA systems from the data

ta_pattern = pattern = r'toxin-antitoxin|addiction module|abortive infection'

ta_names = ['nucleotidyl transferase AbiEii/AbiGii toxin family protein',
            'BrnT family toxin', 
            'zeta toxin family protein']

ta_rows = (seq_id_names['name'].str.contains(ta_pattern, regex=True) |
           seq_id_names['name'].isin(ta_names))

seq_id_names.loc[ta_rows, 'name'].value_counts().head(50)

ta_ids = seq_id_names.loc[ta_rows, 'seq_id'].unique()

### Filtering

background_go_df = go_df.loc[~go_df['seq_id'].isin(true_defense_genes['seq_id']) &
                             ~go_df['seq_id'].isin(defense_go_term_seq_ids) &
                             ~go_df['seq_id'].isin(ta_ids), :]

background_go_df['go_process'].value_counts().head(30)

background_seq_ids = background_go_df['seq_id'].drop_duplicates()

len(background_seq_ids)

background_seq_ids.to_csv('../data3/interim/background_seq_ids.csv', index=False)
background_go_df.to_parquet('../data3/interim/background_go_terms.pq', index=False)
