==> code/04_02_get_secreted_proteins.ipynb <==
from core import read_mmseqs_results
import pandas as pd
import os

seq_id_names = pd.read_parquet('../data/interim/seq_id_names.pq')
go_df = pd.read_parquet('../data/interim/seq_id_go_processes.pq')


conda_commands = ['conda run -n beaker mmseqs']

query_fa = '../data/interim/secreted_proteins.faa'

if 'secreted_proteins.faa' not in os.listdir('../data/interim/'):
    os.system('cat secreted_proteins/*.fasta > ' + query_fa)

working_dir = '../data/interim/mmseqs_working_dir'

query_db = working_dir + 'secreted_proteins_db'

os.sytem(' '.join(conda_commands +
                  ['create_db', 
                   query_fa, 
                   query_db, 
                   '--dbtype 1']))

target_db = working_dir + 'unique_seqs_db'
out_db = working_dir + 'secreted_proteins_refseq_search'

os.system(' '.join(conda_commands +
                   ['search', query_db, target_db, 
                    out_db, '../data/tmp',
                    '--threads 48',
                    '-c 0.8', '--min-seq-id 0.3', 
                    '-a']))

target_out = '../data/interim/secreted_proteins_refseq_search.txt'

os.system(' '.join(conda_commands +
                   ['convertalis', query_db, 
                    target_db, out_db, target_out, 
                    '--threads 48', '--format-output query,target,fident,alnlen,mismatch,gapopen,qstart,qend,tstart,tend,evalue,bits,qcov,tcov
                    ']))

secreted_protein_result = read_mmseqs_results('../data/interim/secreted_proteins_refseq_search.txt')

gold_standard_homologs = secreted_protein_result[(secreted_protein_result['fident'] > 0.9) &
                                                 (secreted_protein_result['qcov'] > 0.9) & 
                                                 (secreted_protein_result['tcov'] > 0.9)]

unique_homologs = gold_standard_homologs['target'].drop_duplicates()

len(unique_homologs)

secreted_protein_rows = seq_id_names['seq_id'].isin(unique_homologs)

seq_id_names.loc[secreted_protein_rows, 'name'].value_counts().head(50)

go_rows = go_df['seq_id'].isin(unique_homologs)

go_df.loc[go_df['seq_id'].isin(unique_homologs), 'go_process'].value_counts()

unique_homologs.to_csv('../data/interim/secreted_homologs.csv', index=False)
