==> code/05_01_mmseqs_candidate.ipynb <==
import os

candidate_seq_f = '../data3/interim/candidate_seqs.faa'

search_out = '../data3/interim/candiate_seq_mmseqs_self_search.txt'

%%time
os.system(' '.join(['conda run -n beaker', 
                    'mmseqs easy-search', 
                    candidate_seq_f, 
                    candidate_seq_f, 
                    search_out,
                    '$TMPDIR',
                    '--format-output query,target,fident,alnlen,mismatch,gapopen,qstart,qend,tstart,tend,evalue,bits,qcov,tcov', 
                    '--threads 48',
                    '-s 7.5']))
