==> code/07_04_cv_foldseek.ipynb <==
import pandas as pd

%%time
foldseek_results = pd.read_table('../data3/interim/foldseek_model_search.txt', 
                                 names=['query','target','fident','alnlen','mismatch','gapopen','qstart','qend','tstart','tend','evalue','bits'])

len(foldseek_results)

**Query:** train gene

**Target:** test gene

model_seq_info = pd.read_parquet('../data3/interim/model_seq_info.pq')

foldseek_results.head()

%%time
top_train_list = list()
for fold, fold_test_df in model_seq_info.groupby('test_fold'):
    print(fold)
    test_set_ids = set(fold_test_df['seq_id'])
    top_foldseek_df = (foldseek_results[~foldseek_results['query'].isin(test_set_ids) & 
                                        foldseek_results['target'].isin(test_set_ids)]
                     .sort_values('bits', ascending=False)
                     .groupby('target')
                     .head(1))
    top_train_list.append(top_foldseek_df)

cat_top_hit = pd.concat(top_train_list)

merged_cat_top_hit = (cat_top_hit[['query', 'target', 'bits']]
                      .merge(model_seq_info[['seq_id', 'defensive']]
                             .rename(columns={'seq_id': 'query', 
                                              'defensive': 'query_defensive'}), 
                             how='inner', on='query'))
merged_cat_top_hit['direction'] = merged_cat_top_hit['query_defensive'] * 2 - 1
merged_cat_top_hit['prediction'] = merged_cat_top_hit['bits'] * merged_cat_top_hit['direction']

out_df = (model_seq_info[['seq_id']]
          .merge(merged_cat_top_hit[['target', 'prediction']]
                 .rename(columns={'target': 'seq_id'}), 
                 how='left', on='seq_id'))
out_df['prediction'] = out_df['prediction'].fillna(0)
out_df['method'] = 'Foldseek/ProstT5 nearest neighbor'

len(out_df)

out_df.to_parquet('../data3/interim/cv_predictions_foldseek.pq', index=False)
