==> code/08_03_cross_validate_gba.ipynb <==
import pandas as pd
from scipy import stats
from joblib import Parallel, delayed
from tqdm import tqdm

def get_fisher_stat(row, total_genes=68_258_704, defensive_genes=244_022):
    random_prob = 1 - (1-defensive_genes/total_genes)**20
    random_defense_total = random_prob*total_genes
    fisher = stats.fisher_exact([[row['defense_count'], row['background_count'] - row['defense_count']], 
                                 [random_defense_total, total_genes - random_defense_total]], alternative='greater')
    return fisher[1]

model_seq_info = pd.read_parquet('../data3/interim/model_seq_info.pq')
model_cluster_assemblies = (pd.read_parquet('../data3/interim/model_cluster_assemblies.pq')
                            .reset_index())
model_cluster_neighbors = pd.read_parquet('../data3/interim/model_cluster_defense_neighbors.pq')

cluster_background_stats = (model_cluster_assemblies.groupby('cluster_id')
                            .agg(background_count=('assembly_stub', 'nunique')))

results_list = list()
for fold, fold_df in model_seq_info.groupby('test_fold'):
    print(fold)
    fold_test_defense_genes = fold_df['defense_gene'].unique()
    filtered_cluster_defense_neighbors = (model_cluster_neighbors
                                          .loc[(~model_cluster_neighbors['defense_neighbor'].isin(fold_test_defense_genes) & 
                                                model_cluster_neighbors['cluster_id'].isin(fold_df['seq_id'])), 
                                               ['seq_id', 'cluster_id', 'assembly']]
                                          .drop_duplicates())
    cluster_defense_stats = (filtered_cluster_defense_neighbors.groupby('cluster_id')
                                     .agg(defense_count=('assembly', 'nunique')))
    fold_clusters = (fold_df[['seq_id']])
    fold_cluster_enrichment = (fold_clusters
                           .merge(cluster_background_stats, left_on='seq_id', 
                                  right_index=True, how='inner')
                           .merge(cluster_defense_stats, left_on='seq_id', 
                                  right_index=True, how='left'))
    fold_cluster_enrichment['defense_count'] = fold_cluster_enrichment['defense_count'].fillna(0)
    fold_cluster_enrichment['frac_defensive'] = (fold_cluster_enrichment['defense_count']/
                                                 fold_cluster_enrichment['background_count'])
    fold_cluster_enrichment['fisher_p'] = Parallel(n_jobs=48)(delayed(get_fisher_stat)(row) for _, row in 
                                                         tqdm(fold_cluster_enrichment.iterrows(), 
                                                              total=len(fold_cluster_enrichment), 
                                                              position=0))
    results_list.append(fold_cluster_enrichment)

results_df = pd.concat(results_list)
results_df['one_minus_fisher_p'] = 1 - results_df['fisher_p']

results_df

out_df = (results_df.rename(columns={'one_minus_fisher_p': 'Guilt-by-association (p-value)', 
                                     'frac_defensive': 'Guilt-by-association (frequency)'})
          .melt(id_vars='seq_id', value_vars=['Guilt-by-association (p-value)', 'Guilt-by-association (frequency)'], 
                value_name='prediction', var_name='method'))

out_df.to_parquet('../data3/interim/cv_predictions_gba.pq', index=False)
