==> code/11_04_build_model_matrix.ipynb <==
import pandas as pd
from joblib import Parallel, delayed
from tqdm import tqdm
import numpy as np

def get_gene_dist(center_seq_id, context_df):
    center_strand = context_df.loc[(context_df['relative_position'] == 0), 'strand'].item()
    if center_strand == '+':
        context_df = context_df.sort_values('relative_position', ascending=True)
    else:
        context_df = context_df.sort_values('relative_position', ascending=False)
    curr_end = context_df['end']
    next_start = context_df['start'].shift(-1)
    if (context_df['end'] < context_df['start']).any():
        context_df['wraparound'] = context_df['end'] < context_df['start']
        next_wraparound = context_df['wraparound'].shift(-1)
        distances = np.where(next_wraparound, 
                             -curr_end,
                             next_start - curr_end)
    else:
        distances = next_start - curr_end
    distances = list(distances)
    out_dict = {'center_seq_id': center_seq_id}
    relative_positions = context_df['relative_position'].to_list()
    for i in range(len(context_df) - 1):
        pos_i = relative_positions[i]
        pos_j = relative_positions[i+1]
        out_dict['dist_' + 
                 ':'.join([str(min(pos_i, pos_j)), 
                           str(max(pos_i, pos_j))])] = distances[i]
    return out_dict

n_feats = 640
feat_names = ['ft' + str(i+1) for i in range(n_feats)]

model_protein_neighbors = pd.read_parquet('../data3/interim/model_neighbors_seq_ids.pq')
model_rep_df = pd.read_parquet('../data3/interim/model_neighbor_representations.pq')
model_rep_df.columns = feat_names

%%time
gene_context_info = pd.read_parquet('../data/interim/refseq_gc_content_breif.pq')

## Get representation features

wide_rep_df = (model_protein_neighbors[['relative_position', 'center_seq_id', 'seq_id']]
                 .merge(model_rep_df, how='inner', 
                        left_on='seq_id', right_index=True)
                 .pivot(index='center_seq_id', columns='relative_position', 
                        values=feat_names))
wide_rep_df = wide_rep_df.fillna(0)
wide_rep_df.columns = [x[0] + '_' + str(x[1]) for x in wide_rep_df.columns]

del model_rep_df

## Get nt features

### Feature class 1: nucleotide

gene_context_info['protein_context_id'] = (gene_context_info['protein_id'] + '|' + 
                                           gene_context_info['genomic_locus'] + '|' +
                                           gene_context_info['start'].astype(str) + '|' +
                                           gene_context_info['strand'])

model_unique_context_ids = model_protein_neighbors['protein_context_id'].drop_duplicates()

nt_features = ['scaled_gc_frac', 'scaled_A_frac', 'scaled_C_frac', 'scaled_T_frac',
               'scaled_G_frac', 'scaled_AA_frac', 'scaled_AC_frac', 'scaled_AT_frac',
               'scaled_AG_frac', 'scaled_CA_frac', 'scaled_CC_frac', 'scaled_CT_frac',
               'scaled_CG_frac', 'scaled_TA_frac', 'scaled_TC_frac', 'scaled_TT_frac', 
               'scaled_TG_frac', 'scaled_GA_frac', 'scaled_GC_frac', 'scaled_GT_frac',
               'scaled_GG_frac']

filtered_gene_context_info = gene_context_info[gene_context_info['protein_context_id'].isin(model_unique_context_ids)]

del gene_context_info

(~model_unique_context_ids.isin(filtered_gene_context_info['protein_context_id'])).sum()

merged_model_protein_neighbors = (model_protein_neighbors
                                  .merge(filtered_gene_context_info[['protein_context_id'] + nt_features], 
                                         how='inner', on='protein_context_id'))

wide_nt_df = merged_model_protein_neighbors.pivot(index='center_seq_id', values=nt_features, columns='relative_position')
wide_nt_df.columns = [x[0] + '_' + str(x[1]) for x in wide_nt_df.columns]
wide_nt_df = wide_nt_df.fillna(1.1)

### Feature class 2: len

len_df = merged_model_protein_neighbors.copy()
len_df['len'] = len_df['seq'].str.len()


wide_len_df = len_df.pivot(index='center_seq_id', values=['len'], columns='relative_position')
wide_len_df.columns = [x[0] + '_' + str(x[1]) for x in wide_len_df.columns]
wide_len_df = wide_len_df.fillna(0)

### Feature class 3: strand

center_strand = merged_model_protein_neighbors.loc[merged_model_protein_neighbors['relative_position'] == 0, 
                                                   ['center_seq_id', 'strand']].rename(columns={'strand': 'center_strand'})
strand_df = merged_model_protein_neighbors.merge(center_strand, how='inner', 
                                                                      on='center_seq_id')
strand_df['co_directional'] = (strand_df['center_strand'] == strand_df['strand']).astype(int)


wide_strand_df = strand_df.pivot(index='center_seq_id', values=['co_directional'], columns='relative_position')
wide_strand_df.columns = [x[0] + '_' + str(x[1]) for x in wide_strand_df.columns]
wide_strand_df = wide_strand_df.fillna(2)

### Feature class 4: distance

protein_dist_list = Parallel(n_jobs=48)(delayed(get_gene_dist)(center_seq_id, context_df) for center_seq_id, context_df in 
                                        tqdm(model_protein_neighbors.groupby('center_seq_id'), 
                                             total=model_protein_neighbors['center_seq_id'].nunique(), 
                                             position=0))

protein_dist_df = pd.DataFrame(protein_dist_list)

protein_dist_df.drop(columns='center_seq_id').min(axis=0)

protein_dist_df = (protein_dist_df.fillna(-200)
                   .set_index('center_seq_id'))

## Merge features

feature_df = wide_rep_df
for df in [wide_nt_df, wide_len_df, wide_strand_df, protein_dist_df]:
    feature_df = (feature_df.merge(df, left_index=True, right_index=True, how='inner'))

feature_df.shape

feature_df

feature_df.to_parquet('../data3/interim/defense_predictor_full_ft_mat.pq', index=True)
