==> code/15_13_analyze_screen_hits.ipynb <==
import pandas as pd
from core import med_log_odds_cutoff

system_info = pd.read_csv('../data3/interim/all_library_info.csv')
display_name_df = pd.read_csv('../data/raw/Beaker validation domains - system_display_names.csv')
ecor_predictions = pd.read_csv('../data3/interim/ecor_predictions_defensive_cat2.csv')

multi_gene_systems = system_info[system_info['max_prediction'] != system_info['min_prediction']]
print('# Multi-gene systems:', len(multi_gene_systems))
print('# With all LO > 0:', (multi_gene_systems['min_prediction'] > 0).sum())

validated_multi_gene_systems = system_info[(system_info['max_prediction'] != system_info['min_prediction']) & 
                                           (system_info['defensive'])]
print('# Multi-gene systems', len(validated_multi_gene_systems))
print('# With all LO > 0:', (validated_multi_gene_systems['min_prediction'] > 0).sum())

validated_multi_gene_systems[validated_multi_gene_systems['min_prediction'] < 0]

13/31

18/31

ecor_seq_assemblies = pd.read_csv('../data/interim/ecor_seq_assemblies.csv', names=['product_accession', 'seq_id', 'seq', 'assembly'])
ecor_seqs = ecor_seq_assemblies[['product_accession', 'seq']].drop_duplicates()

defensive_systems = system_info[system_info['defensive']]
len(defensive_systems)

merged_defensive_systems = (defensive_systems.merge(display_name_df, how='inner', on='working_id')
                            [['working_id', 'DS_name', 'max_prediction']])
merged_defensive_systems['DS_num'] = merged_defensive_systems['DS_name'].str.split('-', expand=True)[1].astype(int)

merged_defensive_systems.sort_values('DS_num')

exploded_systems = defensive_systems.copy()
exploded_systems['product_accession'] = exploded_systems['product_accessions'].str.split(', ')
exploded_systems = exploded_systems.explode('product_accession')
exploded_systems = exploded_systems[['product_accession', 'genomic_accession','working_id', 'screened LB', 'fp', 'pBAD', 'library', 'random', 'curated', 'defensive', 'max_prediction', 'category']]

ecor_predictions.merge(exploded_systems, how='inner', on=['product_accession', 'genomic_accession']).shape

merged_predictions = ecor_predictions.merge(exploded_systems, how='inner', on=['product_accession', 'genomic_accession']).reset_index(drop=True)
merged_predictions['fake_start'] = (pd.Series([-1 if (strand == '-') else 1 for strand in merged_predictions['strand']])) * merged_predictions['start']
merged_predictions['protein_num'] = (merged_predictions.groupby('working_id')
                                     ['fake_start']
                                     .rank()
                                     .astype(int))
merged_predictions = (merged_predictions.merge(ecor_seqs, how='inner', on='product_accession')
                      .merge(display_name_df, how='inner', on='working_id'))
merged_predictions['DS_num'] = merged_predictions['DS_name'].str.split('-', expand=True)[1].astype(int)

true_novel_predictions = (merged_predictions[(merged_predictions['q_cov'] == 0) & 
                                             (merged_predictions['foldseek_q_cov'] == 0) &
                                             (merged_predictions['mean_log_odds'] > med_log_odds_cutoff) & 
                                             ~(merged_predictions['defense_homolog'])].sort_values(['DS_num', 'protein_num']))

len(true_novel_predictions)

true_novel_predictions[['DS_name', 'protein_num', 'mean_log_odds']]

pd.set_option('display.max_columns', 100)
merged_predictions

breif_predictions = (merged_predictions[['protein_context_id', 'assembly', 'genomic_accession', 'start', 'end', 'strand',
                                         'product_accession', 'name', 'seq', 'product_length',
                                         'mean_log_odds',
                                         'defense_homolog_names', 'sys_id', 
                                         'protein_num', 'hit_name', 'q_cov', 'Prob', 
                                          'working_id', 'pBAD', 'library']]
                     .rename(columns={'hit_name': 'HHpred_defense_homolog_name', 
                                      'q_cov': 'HHpred_defense_homolog_q_cov', 
                                      'Prob': 'HHpred_defense_homolog_Prob'}))

with open('../data/interim/hit_proteins.faa', 'w') as f:
    for _, row in breif_predictions.iterrows():
        print('>' + row['product_accession'], file=f)
        print(row['seq'], file=f)

breif_predictions.to_csv('../data3/interim/hit_seq_info.csv', index=False)
