==> code/15_14_get_hit_homologs.ipynb <==
import os
import pandas as pd
from tqdm import tqdm

all_library_info = pd.read_csv('../data3/interim/all_library_info.csv')
novel_cluster_preds = pd.read_csv('../data3/interim/ecor_unique_novel_proteins.csv')

ecor_seq_assemblies = pd.read_csv('../data/interim/ecor_seq_assemblies.csv', names=['product_accession', 'seq_id', 'seq', 'assembly'])
ecor_seqs = ecor_seq_assemblies[['product_accession', 'seq']].drop_duplicates()

long_library_info = all_library_info.copy()
long_library_info['product_accession'] = long_library_info['product_accessions'].str.split(', ')
long_library_info = long_library_info.explode('product_accession')
long_library_info = long_library_info[['product_accession', 'genomic_accession', 'working_id']]
long_library_info = (long_library_info.merge(ecor_seqs, how='inner', on='product_accession'))

with open('../data3/interim/ecor_cloned_seqs.faa', 'w') as f:
    for _, row in long_library_info.iterrows():
        print('>' + row['product_accession'], file=f)
        print(row['seq'], file=f)

cloned_accessions = long_library_info['product_accession'].to_list()

## Run HHPred

faa_file = '../data3/interim/ecor_cloned_seqs.faa'
msa_out_dir = '../data3/interim/ecor_cloned_msas/'
raw_out_dir ='../data3/interim/ecor_cloned_df_alignments/'
parsed_out_file='../data3/interim/ecor_cloned_df_out.csv'

if not os.path.exists(msa_out_dir):
    os.mkdir(msa_out_dir)

old_msa_dir = '../data3/interim/ecor_unique_plus_cloned_msas/'
for msa_f in tqdm(os.listdir(old_msa_dir)):
    if msa_f.split('.a')[0] in cloned_accessions:
        os.system(f"cp {old_msa_dir + msa_f} {msa_out_dir}")

os.system(' '.join(['conda run -n hhpred python', 
                        '~/Documents/hhpred/code/hhpred.py', 
                        '--raw_out_dir', raw_out_dir,  
                        '--parsed_out_file', parsed_out_file, 
                        '--dbs df2', 
                        '--in_fasta', faa_file,  
                        '--n_iter 3', 
                        '--alignment_dir', msa_out_dir, 
                        '--n_jobs 40']))

## Run ESMFold + Foldseek

struct_out_dir = '../data3/interim/ecor_cloned_structures/'
df_rep_struct_db = '../data3/interim/foldseek_dbs/df_rep_foldseek_db_pad'
struct_db = '../data3/interim/foldseek_dbs/ecor_cloned_struct_db'
pad_struct_db = '../data3/interim/foldseek_dbs/ecor_cloned_struct_db_pad'
res_db = '../data3/interim/foldseek_dbs/ecor_cloned_df_rep_search'
struct_out_file = '../data3/interim/ecor_cloned_df_struct_search.txt'
struct_out_html = '../data3/interim/ecor_cloned_df_struct_search.html'

old_struct_dir = '../data3/interim/ecor_validated_structures/'

folded_accs = list()
for struct_f in os.listdir(old_struct_dir):
    if struct_f.split('.p')[0] in cloned_accessions:
        os.system(f"cp {old_struct_dir + struct_f} {struct_out_dir + struct_f}")
        folded_accs.append(struct_f.split('.p')[0])

struct_faa_file = '../data3/interim/ecor_cloned_seqs_missing_fold.faa'

filtered_library_info = long_library_info[~long_library_info['product_accession'].isin(folded_accs)]
with open(struct_faa_file, 'w') as f:
    for _, row in filtered_library_info.iterrows():
        print('>' + row['product_accession'], file=f)
        print(row['seq'], file=f)

os.system(' '.join(['conda run -n esmfold3 python', 
                    'fold.py', 
                    '-i', struct_faa_file, 
                    '-o', struct_out_dir]))

foldseek_preamble = ['conda run -n nvcc /home/gridsan/pdeweirdt/Documents/foldseek/build/bin/foldseek']

os.system(' '.join(foldseek_preamble + 
                   ['createdb', struct_out_dir, struct_db]))

os.system(' '.join(foldseek_preamble +
                   ['makepaddedseqdb', struct_db, pad_struct_db]))

os.system(' '.join(foldseek_preamble +
                   ['search', pad_struct_db, df_rep_struct_db, 
                    res_db, '$TMPDIR', 
                    '--gpu 1', '--alignment-type 1', '-a']))

os.system(' '.join(foldseek_preamble +
                   ['convertalis', pad_struct_db, df_rep_struct_db, 
                    res_db, struct_out_file, 
                    '--format-output', 'query,target,fident,alnlen,mismatch,gapopen,qstart,qend,tstart,tend,qcov,tcov,evalue,bits,alntmscore,qtmscore,ttmscore,rmsd,lddt,prob']))

os.system(' '.join(foldseek_preamble +
                   ['convertalis', pad_struct_db, df_rep_struct_db, 
                    res_db, struct_out_html, 
                    '--format-mode 3']))

## Analyze homology results

hhpred_results = pd.read_csv(parsed_out_file)
foldseek_results = pd.read_table(struct_out_file,
                                 names=["query", "target", "fident", "alnlen", "mismatch", "gapopen", "qstart", "qend", "tstart", "tend", "qcov", "tcov", "evalue", "bits", "alntmscore", "qtmscore", "ttmscore", "rmsd", "lddt", "prob"])
predictions = pd.read_csv('../data3/interim/ecor_predictions_defensive_cat2.csv')
hit_seq_info = pd.read_csv('../data3/interim/hit_seq_info.csv')
display_name_df = pd.read_csv('../data/raw/Beaker validation domains - system_display_names.csv')
df_rep_info = pd.read_csv('../data/interim/df_top_uniprot_hits.csv')

hit_categories = ['Defense homolog in expected system',
                  'Defense homolog in new context', 
                  'Remote defense homolog', 
                  'Structural defense homolog',
                  'Predicted novel defense gene', 
                  'Not defensive']
predictions['hit_category'] = pd.Categorical(predictions['hit_category'], 
                                             categories=hit_categories)

non_redundant_predictions = (predictions
                             .sort_values(['hit_category', 'mean_log_odds'], ascending=[True, False])
                             .groupby('cluster_id')
                             .head(1))

hhpred_results['q_ali_len'] = hhpred_results['qend'] - hhpred_results['qstart']
hhpred_homologs = (hhpred_results[(hhpred_results['Prob'] > 50)]
                           .sort_values('q_ali_len', ascending=False)
                           .groupby('query')
                           .head(1))

foldseek_homologs = (foldseek_results[foldseek_results['prob'] > 0.6]
                     .sort_values('prob', ascending=False)
                     .groupby('query')
                     .head(1))

**Assigning validated hits**

If hit is part of cluster that is not not defensive:

    Assign hit to category

Elif hit has log-odds > 0:
    
    Assign hit to category based on hhpred and foldseek results

Else:

    Assign hit to non-defensive category


def assign_hit_category(row, qcov=0.4):
    if row['mean_log_odds'] < 0:
        return 'Not defensive'
    elif row['cluster_hit_category'] != 'Not defensive':
        return row['cluster_hit_category']
    elif row['defense_system_protein']:
        return 'Defense homolog in expected system'
    elif row['defense_homolog']:
        return 'Defense homolog in new context'
    elif row['hhpred_qcov'] > qcov:
        return 'Remote defense homolog'
    elif row['foldseek_qcov'] > qcov:
        return 'Structural defense homolog'
    else:
        return 'Predicted novel defense gene'

merged_long_library_info = (long_library_info.merge(predictions[['product_accession', 'genomic_accession','cluster_id', 'defense_homolog', 
                                                                 'defense_system_protein', 'product_length', 'mean_log_odds', 'start', 'strand']], 
                                          how='inner', on=['product_accession', 'genomic_accession'])
                       .merge(non_redundant_predictions[['cluster_id', 'hit_category']]
                              .rename(columns={'hit_category': 'cluster_hit_category'}), 
                              how='inner', on='cluster_id')
                       .merge(display_name_df, how='left', on='working_id')
                       .merge(hhpred_homologs[['query', 'hit_name', 'q_ali_len']]
                              .rename(columns={'query': 'product_accession', 
                                               'hit_name': 'hhpred_hit_name', 
                                               'q_ali_len': 'hhpred_q_ali_len'}), 
                              how='left', on='product_accession')
                       .merge(foldseek_homologs[['query', 'target', 'qcov']]
                              .merge(df_rep_info[['target', 'description']], how='inner', on='target')
                              .rename(columns={'query': 'product_accession',
                                               'description': 'foldseek_target_homolog',
                                               'target': 'foldseek_target',
                                               'qcov': 'foldseek_qcov'}), 
                              how='left', on='product_accession'))
merged_long_library_info['fake_start'] = pd.Series([1 if x == '+' else -1 for x in merged_long_library_info['strand']]) * merged_long_library_info['start']
merged_long_library_info['protein_num'] = merged_long_library_info.groupby('working_id')['fake_start'].rank()
merged_long_library_info['hhpred_qcov'] = merged_long_library_info['hhpred_q_ali_len']/merged_long_library_info['product_length']
merged_long_library_info['hit_category'] = merged_long_library_info.apply(assign_hit_category, axis=1)
merged_long_library_info['hit_category'] = pd.Categorical(merged_long_library_info['hit_category'], categories=['Not defensive'] + hit_categories[:-1])
id_top_category = (merged_long_library_info.sort_values('hit_category', ascending=False)
                   .groupby('working_id')
                   .head(1)
                   [['working_id', 'hit_category', 'DS_name']]
                   .rename(columns={'hit_category':'top_hit_category'}))
merged_long_library_info = (merged_long_library_info.merge(id_top_category.drop(columns='DS_name'), how='inner', on='working_id'))

len(merged_long_library_info)

novel_cluster_preds.columns

no_cov_novel_clusters = novel_cluster_preds.loc[novel_cluster_preds['q_cov'] == 0, 'product_accession']

no_cov_validated_clusters = (merged_long_library_info[~merged_long_library_info['DS_name'].isna() & 
                                                      merged_long_library_info['cluster_id'].isin(no_cov_novel_clusters)]
                             .reset_index(drop=True))
no_cov_validated_clusters['ds_num'] = no_cov_validated_clusters['DS_name'].str.split('-', expand=True)[1].astype(int)
print('# clusters:', no_cov_validated_clusters['cluster_id'].nunique())
print('# proteins', len(no_cov_validated_clusters))

no_cov_validated_clusters.sort_values(['ds_num', 'protein_num'])[['DS_name', 'protein_num', 'hit_category']]

(~merged_long_library_info['DS_name'].isna() & (merged_long_library_info['cluster_id'].isin(no_cov_novel_clusters)))

((merged_long_library_info['hit_category'] == 'Predicted novel defense gene') & 
 merged_long_library_info['hhpred_q_ali_len'].isna() & 
 ~merged_long_library_info['DS_name'].isna()).sum()


id_top_category['top_hit_category'].value_counts()

id_top_category.loc[~id_top_category['DS_name'].isna(), 'top_hit_category'].value_counts()

highlight_hit_seq_info = merged_long_library_info[~merged_long_library_info['DS_name'].isna()].copy()
highlight_hit_seq_info['ds_num'] = highlight_hit_seq_info['DS_name'].str.split('-', expand=True)[1].astype(int)
highlight_hit_seq_info = (highlight_hit_seq_info
                          .sort_values(['ds_num', 'protein_num'])
                          .loc[highlight_hit_seq_info['hit_category'] == highlight_hit_seq_info['top_hit_category'], 
                                                  ['DS_name', 'protein_num', 'hit_category']])

highlight_hit_seq_info

id_top_category['top_hit_category'].value_counts()
