==> code/15_17_get_potential_negatives.ipynb <==
import pandas as pd
import os

all_library_info = pd.read_csv('../data3/interim/all_library_info.csv')
v1_neg_info = pd.read_csv('../data/raw/Beaker rebuttal domains - hit_proteins.csv')
ecor_predictions = pd.read_csv('../data3/interim/ecor_predictions.csv')

top_preds = (ecor_predictions.sort_values('mean_log_odds', ascending=False)
             .groupby('cluster_id')
             .head(1))

len(top_preds)

cloned_preds = (ecor_predictions[ecor_predictions['product_accession']
                                 .isin(all_library_info['product_accessions'].str.split(', ').explode()) |
                                 ecor_predictions['product_accession']
                                 .isin(v1_neg_info['product_accession'])])

'WP_016240615.1' in cloned_preds['product_accession'].to_list()

'WP_016240615.1' in all_library_info.loc[all_library_info['working_id'] == 'DISA', 'product_accessions'].str.split(', ').explode().to_list()

all_library_info[all_library_info['working_id'] == 'DISA']

cat_preds = (pd.concat([top_preds, 
                       cloned_preds])
             .drop_duplicates())

len(cat_preds)

unique_cat_pred_accs = cat_preds[['product_accession']].drop_duplicates()

len(unique_cat_pred_accs)

id_f = '../data3/interim/ecor_unique_plus_cloned_ids.txt'

unique_cat_pred_accs.to_csv(id_f, index=False, header=False)

faa_f = '../data3/interim/ecor_unique_plus_cloned_ids.faa'

os.system(' '.join(['conda run -n beaker', 
                    'seqtk subseq',
                    '../data/interim/ecor_unique_seqs.faa', 
                        id_f, '>',
                        faa_f]))
