==> code/18_00_test_new_df_padloc_hmms.ipynb <==
import os
import pandas as pd
from core import read_dom_table

padloc_hmm_info = pd.read_table('https://raw.githubusercontent.com/padlocbio/padloc-db/refs/heads/master/hmm_meta.txt')

## Set up hmm databases

#padloc_db = '/home/gridsan/pdeweirdt/.conda/envs/padloc/data/hmm/padlocdb.hmm'
df_db = '../data3/interim/defense_finder_db.hmm'

top_dir = '/home/deweirdt/.macsyfinder/models/CasFinder/profiles/'
cas_finder_hmms = [os.path.join(top_dir, x) for x in os.listdir(top_dir) if x.endswith('.hmm') and x not in {'cas12a_V-A_4.hmm', 'cas4_V_1.hmm', 'cas9_II-B_3.hmm'}]

if 'defense_finder_db.hmm' not in os.listdir('../data3/interim/'):
    os.system(f"cat {' '.join(cas_finder_hmms)} /home/deweirdt/.macsyfinder/models/defense-finder-models/profiles/*.hmm > {df_db}")

## Seach E. coli 3000

ecoli_3k_novel_df = pd.read_parquet('../data3/interim/ecoli3k_predicted_defense_struct.pq')
ecoli_3k_novel_df = (ecoli_3k_novel_df[ecoli_3k_novel_df['hit_category'] == 'Predicted novel defense gene']
                     .reset_index(drop=True))
len(ecoli_3k_novel_df)

ecoli_3k_novel_ids = '../data3/interim/ecoli_novel_ids.txt'

ecoli_3k_novel_df[['product_accession']].to_csv(ecoli_3k_novel_ids, header=False, index=False)

ecoli_3k_seqs = '../data/ecoli/interim/unique_seqs.faa'

ecoli_3k_novel_seqs = '../data3/interim/ecoli_3k_unique_novel_seqs.faa'

!conda run -n beaker seqtk subseq {ecoli_3k_seqs} {ecoli_3k_novel_ids} > {ecoli_3k_novel_seqs}

ecoli_3k_padloc_out = '../data3/interim/ecoli_3k_padloc_domtbl.txt'
ecoli_3k_padloc_temp_out = '../data/tmp/temp_ecoli_3k'

!conda run -n beaker hmmsearch --cpu 40 --acc --noali --domtblout {ecoli_3k_padloc_out} -o {ecoli_3k_padloc_temp_out} {padloc_db} {ecoli_3k_novel_seqs}


ecoli_3k_df_out = '../data3/interim/ecoli_3k_df_domtbl.txt'
ecoli_3k_df_temp_out = '../data/tmp/temp_ecoli_3k'

!conda run -n defensefinder hmmsearch --cpu 40 --acc --noali --domtblout {ecoli_3k_df_out} -o {ecoli_3k_df_temp_out} {df_db} {ecoli_3k_novel_seqs}


## Analyze hits

ecoli_3k_padloc_domtbl = read_dom_table(ecoli_3k_padloc_out)
ecoli_3k_padloc_domtbl['hmm_db'] = 'padloc'
ecoli_3k_padloc_domtbl['seq_db'] = 'ecoli_3k'
ecoli_3k_df_domtbl = read_dom_table(ecoli_3k_df_out)
ecoli_3k_df_domtbl['hmm_db'] = 'df'
ecoli_3k_df_domtbl['seq_db'] = 'ecoli_3k'

ecoli_3k_df_domtbl['target'].nunique()

for dom_df in [ecoli_3k_padloc_domtbl, ecoli_3k_df_domtbl]:
    dom_df['hmm_ali_len'] = dom_df['hmm_to'] - dom_df['hmm_from'] + 1
    dom_df['env_ali_len'] = dom_df['env_to'] - dom_df['env_from'] + 1
    dom_df['hmm_cov'] = dom_df['hmm_ali_len'] / dom_df['qlen']
    dom_df['t_cov'] = dom_df['env_ali_len'] / dom_df['tlen']


merged_ecoli_3k_padloc_domtbl = (ecoli_3k_padloc_domtbl.merge(padloc_hmm_info[['hmm.accession', 'e.val.threshold', 
                                                                                   'hmm.coverage.threshold', 'target.coverage.threshold']]
                                                                  .rename(columns={'hmm.accession': 'query_accession'}), 
                                                                  how='inner', 
                                                                  on='query_accession'))

merged_ecoli_3k_padloc_domtbl[['hmm.coverage.threshold', 'target.coverage.threshold', 'hmm_cov', 't_cov', 'dom_i_evalue', 'e.val.threshold']].agg(['mean','median', 'std', 'max', 'min'])

filtered_ecoli_3k_padloc_domtbl = merged_ecoli_3k_padloc_domtbl[(merged_ecoli_3k_padloc_domtbl['dom_i_evalue'] < merged_ecoli_3k_padloc_domtbl['e.val.threshold']) & 
                                                                    (merged_ecoli_3k_padloc_domtbl['hmm_cov'] > merged_ecoli_3k_padloc_domtbl['hmm.coverage.threshold']) &
                                                                    (merged_ecoli_3k_padloc_domtbl['t_cov'] > merged_ecoli_3k_padloc_domtbl['target.coverage.threshold']) & 
                                                                    (~merged_ecoli_3k_padloc_domtbl['query'].str.startswith('PDC'))]
filtered_ecoli_3k_padloc_domtbl['target'].nunique()

(filtered_ecoli_3k_padloc_domtbl.sort_values('dom_i_evalue')
 .groupby('query')
 .head(1))

df_i_eval = 0.001
df_hmm_cov = 0.4

filtered_ecoli_3k_df_domtbl = ecoli_3k_df_domtbl[(ecoli_3k_df_domtbl['dom_i_evalue'] < df_i_eval) & 
                                                     (ecoli_3k_df_domtbl['hmm_cov'] > df_hmm_cov)]
filtered_ecoli_3k_df_domtbl['target'].nunique()

top_ecoli_3k = (filtered_ecoli_3k_df_domtbl.sort_values('dom_i_evalue')
                .groupby('target')
                .head(1))

len(top_ecoli_3k)

top_ecoli_3k['query'].str.contains('DS').sum()

top_ecoli_3k['query'].value_counts()

200/1041
